requirements.txt
ipython
python-dotenv
requests
tqdm
